src/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
pyjwt==2.8.0
cryptography==41.0.7
sqlalchemy==2.0.44
bcrypt==4.1.2
pymysql==1.1.0
psycopg2-binary==2.9.9
flask==3.0.0
requests==2.31.0